+++ schemacrawler/src/main/java/schemacrawler/tools/operation/OperatorPlainTextOutput.java
@@ -60,7 +60,7 @@ final class OperatorPlainTextOutput
    *      java.lang.String, java.lang.String, int, java.lang.String)
    */
   public void handleTable(final int ordinalPosition, final String tableName,
-                          final String tableType, final int count,
+                          final String tableType, final long count,
                           final String message)
   {
     out.print(Utilities.padRight(tableName, MAX_TABLE_NAME_WIDTH));
